refactoring games into utils v api
